refactor: extract classes
