+++ src/test/java/org/cicirello/search/problems/LCSTests.java
@@ -22,7 +22,6 @@ package org.cicirello.search.problems;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.util.ArrayList;
 import java.util.Iterator;
 import org.cicirello.permutations.Permutation;
 import org.junit.jupiter.api.*;
@@ -120,263 +119,6 @@ public class LCSTests {
         assertThrows(
             IllegalArgumentException.class,
             () -> new LargestCommonSubgraph(5, 5, 0.5, 1 + 1E-10, 42));
-  }
-
-  @Test
-  public void testPetersenGraphInstance() {
-    for (int n = 1; n <= 7; n++) {
-      for (int k = 0; k < 0.5 * n; k++) {
-        LargestCommonSubgraph lcs =
-            LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(n, k);
-        assertEquals(2 * n, lcs.size());
-        assertEquals(3 * n, lcs.maxValue());
-        for (int i = 0; i < n; i++) {
-          assertTrue(lcs.hasEdge1(i, (i + 1) % n));
-          assertTrue(lcs.hasEdge1(n + i, n + ((i + k) % n)));
-          assertTrue(lcs.hasEdge1(i, n + i));
-        }
-        lcs = LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(n, k, 42);
-        assertEquals(2 * n, lcs.size());
-        assertEquals(3 * n, lcs.maxValue());
-        for (int i = 0; i < n; i++) {
-          assertTrue(lcs.hasEdge1(i, (i + 1) % n));
-          assertTrue(lcs.hasEdge1(n + i, n + ((i + k) % n)));
-          assertTrue(lcs.hasEdge1(i, n + i));
-        }
-      }
-    }
-    IllegalArgumentException thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(1, 1));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(2, 1));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(3, 2));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(4, 2));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(5, 3));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(6, 3));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(1, 1, 42));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(2, 1, 42));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(3, 2, 42));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(4, 2, 42));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(5, 3, 42));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(6, 3, 42));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(5, -1));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(5, -1, 42));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(0, 0));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class,
-            () -> LargestCommonSubgraph.createInstanceGeneralizedPetersenGraph(0, 0, 42));
-  }
-
-  @Test
-  public void testSpecificGraphsCase() {
-    LargestCommonSubgraph.Edge edge = new LargestCommonSubgraph.Edge(5, 7);
-    assertEquals(5, edge.getU());
-    assertEquals(7, edge.getV());
-
-    ArrayList<LargestCommonSubgraph.Edge> edges1 = new ArrayList<LargestCommonSubgraph.Edge>();
-    edges1.add(new LargestCommonSubgraph.Edge(0, 2));
-    edges1.add(new LargestCommonSubgraph.Edge(0, 4));
-    edges1.add(new LargestCommonSubgraph.Edge(1, 2));
-    edges1.add(new LargestCommonSubgraph.Edge(3, 0));
-    int[][] in1 = {{0, 2}, {0, 4}, {1, 2}, {3, 0}, {2, 0}, {4, 0}, {2, 1}, {0, 3}};
-    int[][] notIn1 = {
-      {0, 1}, {1, 0}, {1, 3}, {1, 4}, {2, 3}, {2, 4}, {3, 1}, {3, 2}, {3, 4}, {4, 1}, {4, 2}, {4, 3}
-    };
-    ArrayList<LargestCommonSubgraph.Edge> edges2 = new ArrayList<LargestCommonSubgraph.Edge>();
-    edges2.add(new LargestCommonSubgraph.Edge(0, 3));
-    edges2.add(new LargestCommonSubgraph.Edge(4, 3));
-    edges2.add(new LargestCommonSubgraph.Edge(4, 1));
-    edges2.add(new LargestCommonSubgraph.Edge(1, 3));
-    edges2.add(new LargestCommonSubgraph.Edge(0, 1));
-    int[][] in2 = {{0, 3}, {3, 4}, {4, 1}, {1, 3}, {0, 1}, {3, 0}, {4, 3}, {1, 4}, {3, 1}, {1, 0}};
-    int[][] notIn2 = {
-      {0, 2}, {0, 4}, {1, 2}, {2, 3}, {2, 4}, {2, 0}, {2, 1}, {3, 2}, {4, 0}, {4, 2}
-    };
-
-    LargestCommonSubgraph lcs = new LargestCommonSubgraph(5, 5, edges1, edges2);
-    for (int[] e : in1) {
-      assertTrue(lcs.hasEdge1(e[0], e[1]));
-      assertTrue(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : in2) {
-      assertTrue(lcs.hasEdge2(e[0], e[1]));
-      assertTrue(lcs.hasEdge2(e[1], e[0]));
-    }
-    for (int[] e : notIn1) {
-      assertFalse(lcs.hasEdge1(e[0], e[1]));
-      assertFalse(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : notIn2) {
-      assertFalse(lcs.hasEdge2(e[0], e[1]));
-      assertFalse(lcs.hasEdge2(e[1], e[0]));
-    }
-    assertEquals(4, lcs.maxValue());
-    assertEquals(5, lcs.size());
-
-    lcs = new LargestCommonSubgraph(5, 5, edges2, edges1);
-    for (int[] e : in1) {
-      assertTrue(lcs.hasEdge1(e[0], e[1]));
-      assertTrue(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : in2) {
-      assertTrue(lcs.hasEdge2(e[0], e[1]));
-      assertTrue(lcs.hasEdge2(e[1], e[0]));
-    }
-    for (int[] e : notIn1) {
-      assertFalse(lcs.hasEdge1(e[0], e[1]));
-      assertFalse(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : notIn2) {
-      assertFalse(lcs.hasEdge2(e[0], e[1]));
-      assertFalse(lcs.hasEdge2(e[1], e[0]));
-    }
-    assertEquals(4, lcs.maxValue());
-    assertEquals(5, lcs.size());
-
-    lcs = new LargestCommonSubgraph(5, 6, edges1, edges2);
-    for (int[] e : in1) {
-      assertTrue(lcs.hasEdge1(e[0], e[1]));
-      assertTrue(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : in2) {
-      assertTrue(lcs.hasEdge2(e[0], e[1]));
-      assertTrue(lcs.hasEdge2(e[1], e[0]));
-    }
-    for (int[] e : notIn1) {
-      assertFalse(lcs.hasEdge1(e[0], e[1]));
-      assertFalse(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : notIn2) {
-      assertFalse(lcs.hasEdge2(e[0], e[1]));
-      assertFalse(lcs.hasEdge2(e[1], e[0]));
-    }
-    assertEquals(4, lcs.maxValue());
-    assertEquals(6, lcs.size());
-
-    lcs = new LargestCommonSubgraph(5, 6, edges2, edges1);
-    for (int[] e : in2) {
-      assertTrue(lcs.hasEdge1(e[0], e[1]));
-      assertTrue(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : in1) {
-      assertTrue(lcs.hasEdge2(e[0], e[1]));
-      assertTrue(lcs.hasEdge2(e[1], e[0]));
-    }
-    for (int[] e : notIn2) {
-      assertFalse(lcs.hasEdge1(e[0], e[1]));
-      assertFalse(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : notIn1) {
-      assertFalse(lcs.hasEdge2(e[0], e[1]));
-      assertFalse(lcs.hasEdge2(e[1], e[0]));
-    }
-    assertEquals(4, lcs.maxValue());
-    assertEquals(6, lcs.size());
-
-    lcs = new LargestCommonSubgraph(6, 5, edges1, edges2);
-    for (int[] e : in2) {
-      assertTrue(lcs.hasEdge1(e[0], e[1]));
-      assertTrue(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : in1) {
-      assertTrue(lcs.hasEdge2(e[0], e[1]));
-      assertTrue(lcs.hasEdge2(e[1], e[0]));
-    }
-    for (int[] e : notIn2) {
-      assertFalse(lcs.hasEdge1(e[0], e[1]));
-      assertFalse(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : notIn1) {
-      assertFalse(lcs.hasEdge2(e[0], e[1]));
-      assertFalse(lcs.hasEdge2(e[1], e[0]));
-    }
-    assertEquals(4, lcs.maxValue());
-    assertEquals(6, lcs.size());
-
-    lcs = new LargestCommonSubgraph(6, 5, edges2, edges1);
-    for (int[] e : in1) {
-      assertTrue(lcs.hasEdge1(e[0], e[1]));
-      assertTrue(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : in2) {
-      assertTrue(lcs.hasEdge2(e[0], e[1]));
-      assertTrue(lcs.hasEdge2(e[1], e[0]));
-    }
-    for (int[] e : notIn1) {
-      assertFalse(lcs.hasEdge1(e[0], e[1]));
-      assertFalse(lcs.hasEdge1(e[1], e[0]));
-    }
-    for (int[] e : notIn2) {
-      assertFalse(lcs.hasEdge2(e[0], e[1]));
-      assertFalse(lcs.hasEdge2(e[1], e[0]));
-    }
-    assertEquals(4, lcs.maxValue());
-    assertEquals(6, lcs.size());
-
-    IllegalArgumentException thrown =
-        assertThrows(
-            IllegalArgumentException.class, () -> new LargestCommonSubgraph(0, 5, edges2, edges1));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class, () -> new LargestCommonSubgraph(5, 0, edges2, edges1));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class, () -> new LargestCommonSubgraph(4, 5, edges1, edges2));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class, () -> new LargestCommonSubgraph(4, 5, edges2, edges1));
-    edges2.add(new LargestCommonSubgraph.Edge(5, 2));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class, () -> new LargestCommonSubgraph(5, 5, edges1, edges2));
-    edges2.set(edges2.size() - 1, new LargestCommonSubgraph.Edge(2, 5));
-    thrown =
-        assertThrows(
-            IllegalArgumentException.class, () -> new LargestCommonSubgraph(5, 5, edges1, edges2));
   }
 
   @Test
